+++ ambient_download/src/cli.rs
@@ -1,5 +1,6 @@
 use clap::{Arg, Command}; // Command line
 
+#[allow(clippy::too_many_lines, clippy::module_name_repetitions)]
 pub fn build_cli() -> Command {
     Command::new(clap::crate_name!())
         .about(clap::crate_description!())
@@ -126,5 +127,15 @@ pub fn build_cli() -> Command {
         )
         .subcommand(Command::new("timezone")
             .about("Print the local time zone information.")
+        )
+        .subcommand(Command::new("newconfig")
+            .about("Create a new, empty configuration file.")
+            .arg( // Config-file
+                Arg::new("config-file")
+                    .value_name("FILENAME")
+                    .help("The name of the configuration file to create.")
+                    .long_help("The configuration file in TOML format to create. Default is 'ambient_download.toml' if not specified.")
+                    .num_args(..=1)
+            )
         )
 }